+++ tests/MethodTest.php
@@ -5,7 +5,8 @@ declare(strict_types=1);
 namespace Barryvdh\LaravelIdeHelper\Tests;
 
 use Barryvdh\LaravelIdeHelper\Method;
-use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Database\Eloquent\Builder as EloquentBuilder;
+use Illuminate\Database\Query\Builder as QueryBuilder;
 use PHPUnit\Framework\TestCase;
 
 class MethodTest extends TestCase
@@ -54,16 +55,16 @@ DOC;
     }
 
     /**
-     * Test the output of a class
+     * Test the output of Illuminate\Database\Eloquent\Builder
      */
     public function testEloquentBuilderOutput()
     {
-        $reflectionClass = new \ReflectionClass(Builder::class);
+        $reflectionClass = new \ReflectionClass(EloquentBuilder::class);
         $reflectionMethod = $reflectionClass->getMethod('with');
 
-        $method = new Method($reflectionMethod, 'Builder', $reflectionClass);
+        $method = new Method($reflectionMethod, 'Builder', $reflectionClass, null, [], [], ['$this' => '\\' . EloquentBuilder::class . '|static']);
 
-        $output =  <<<'DOC'
+        $expectedDocComment =  <<<'DOC'
 /**
  * Set the relationships that should be eager loaded.
  *
@@ -73,14 +74,44 @@ DOC;
  * @static 
  */
 DOC;
-        $this->assertSame($output, $method->getDocComment(''));
+        $this->assertSame($expectedDocComment, $method->getDocComment(''));
         $this->assertSame('with', $method->getName());
-        $this->assertSame('\\' . Builder::class, $method->getDeclaringClass());
-        $this->assertSame('$relations, $callback', $method->getParams(true));
+        $this->assertSame('\\' . EloquentBuilder::class, $method->getDeclaringClass());
         $this->assertSame(['$relations', '$callback'], $method->getParams(false));
-        $this->assertSame('$relations, $callback = null', $method->getParamsWithDefault(true));
         $this->assertSame(['$relations', '$callback = null'], $method->getParamsWithDefault(false));
         $this->assertTrue($method->shouldReturn());
+        $this->assertSame('\Illuminate\Database\Eloquent\Builder|static', rtrim($method->getReturnTag()->getType()));
+    }
+
+    /**
+     * Test the output of Illuminate\Database\Query\Builder
+     */
+    public function testQueryBuilderOutput()
+    {
+        $reflectionClass = new \ReflectionClass(QueryBuilder::class);
+        $reflectionMethod = $reflectionClass->getMethod('whereNull');
+
+        $method = new Method($reflectionMethod, 'Builder', $reflectionClass, null, [], [], ['$this' => '\\' . EloquentBuilder::class . '|static']);
+
+        $expectedDocComment =  <<<'DOC'
+/**
+ * Add a "where null" clause to the query.
+ *
+ * @param string|array|\Illuminate\Contracts\Database\Query\Expression $columns
+ * @param string $boolean
+ * @param bool $not
+ * @return \Illuminate\Database\Eloquent\Builder|static 
+ * @static 
+ */
+DOC;
+
+        $this->assertSame($expectedDocComment, $method->getDocComment(''));
+        $this->assertSame('whereNull', $method->getName());
+        $this->assertSame('\\' . QueryBuilder::class, $method->getDeclaringClass());
+        $this->assertSame(['$columns', '$boolean', '$not'], $method->getParams(false));
+        $this->assertSame(['$columns', "\$boolean = 'and'", '$not = false'], $method->getParamsWithDefault(false));
+        $this->assertTrue($method->shouldReturn());
+        $this->assertSame('\Illuminate\Database\Eloquent\Builder|static', rtrim($method->getReturnTag()->getType()));
     }
 
     /**